+++ Stipstonks/Items/StonkMarketEventProgressItem.cs
@@ -6,6 +6,9 @@ namespace Stip.Stipstonks.Items;
 public partial class StonkMarketEventProgressItem : ObservableObject
 {
     [ObservableProperty]
+    private bool _isVisible = true;
+
+    [ObservableProperty]
     private TimeSpan _duration;
 
     [ObservableProperty]