+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## 2.0.0
 ### Stipstonks
+* An option was added to hide the crash progress bar.
 * The application was ported to `Avalonia UI`.
 ### Beer Battle
 * This is the initial version.